No more "content/" in texture names. Default window is now 1600x900

+++ Dev/StaticSheepNeoBuildSystem/game/SheepGraphics/Context/Context.cpp
@@ -115,7 +115,7 @@ namespace DirectSheep
 
     m_swapChain->SetFullscreenState(FALSE, NULL);
 
-    for (int i = 0; i < m_handles.size(); ++i)
+    for (size_t i = 0; i < m_handles.size(); ++i)
     {
       Release(m_handles[i]);
     }
